Java gstreamer bindings. Added more signals, fixed bus to coalesce signals, and added xoverlay.

+++ trunk/Main/OSSBuild/Java/Libraries/media-gstreamer-bindings/src/ossbuild/media/gstreamer/api/GType.java
@@ -115,7 +115,7 @@ public enum GType {
 	}
 
 	public static boolean addCustomConverter(NativeLong nativeValue, IGTypeConverter converter) {
-		return addCustomConverter(nativeValue, converter);
+		return addCustomConverter(nativeValue.longValue(), converter);
 	}
 
 	public static boolean addCustomConverter(long nativeValue, IGTypeConverter converter) {
